found why items arent adding
staticinterface script and dynamicinterface script clashing in terms of slot creations and when item is picked up it returns no empty slots avabile, so needing to impliment a feature where the item either doesn't get picked up if no slots or the item is dropped on the ground.

removed static interface script, have new bug in mc_controller about whether invenotry is open or not now. haven't tested if items is fixed.